+++ sips/h/fileio.py
@@ -1,3 +1,7 @@
+import os
+import random
+
+
 def init_csv(fn, header, close=True):
     """
 
@@ -20,3 +24,29 @@ def write_list(file, list):
             file.write("\n")
         else:
             file.write(",")
+
+
+def get_fns(dir):
+    fns = os.listdir(dir)
+    try:
+        fns.remove("LOG.csv")
+    except ValueError:
+        pass
+
+    return fns
+
+
+def train_test_split_dir(fns, train_frac=0.7, shuffle=False):
+    """
+
+    """
+    num_files = len(fns)
+    split_idx = round(0.7 * num_files)
+
+    if shuffle:
+        random.shuffle(fns)
+
+    train_fns = fns[0:split_idx]
+    test_fns = fns[split_idx:]
+
+    return train_fns, test_fns
