+++ sips/lines/lines.py
@@ -12,14 +12,20 @@ import json
 import sips
 import sips.h.fileio as io
 import sips.h.helpers as h
+import sips.h.serialize as s
+
 from sips.h.cloud import profiler
+from sips.h import analyze
+from sips.h import hot
+
 from sips.lines import collate
-from sips.macros import nfl
-from sips.macros import nba
-from sips.macros import nhl
-from sips.macros import macros as m
-from sips.macros import bov as bm
 from sips.lines.bov import bov
+
+from sips.macros import macros as m
+from sips.macros.sports import nfl
+from sips.macros.sports import nba
+from sips.macros.sports import nhl
+from sips.macros import bov as bm
 from sips.lines.bov.utils import bov_utils as utils
 
 import concurrent.futures
@@ -32,7 +38,8 @@ CONFIG_PATH = m.PROJ_DIR + "lines/config/lines.json"
 
 
 parser = argparse.ArgumentParser(description="configure lines.py")
-parser.add_argument("-d", "--dir", type=str, help="folder name of run", default="run3")
+parser.add_argument("-d", "--dir", type=str,
+                    help="folder name of run", default="run3")
 group = parser.add_mutually_exclusive_group()
 group.add_argument(
     "-s",
@@ -50,7 +57,8 @@ parser.add_argument("-n", "--new_only", type=bool, help="", default=True)
 parser.add_argument(
     "-w", "--wait", type=float, help="how long to wait after each step", default=0.25
 )
-parser.add_argument("-v", "--verbose", type=bool, help="print more", default=False)
+parser.add_argument("-v", "--verbose", type=bool,
+                    help="print more", default=False)
 parser.add_argument(
     "-c", "--grab_espn", type=bool, help="collate with espn data", default=False
 )
@@ -197,7 +205,8 @@ class Lines:
         if os.path.isfile(self.log_path):
             self.log_file = open(self.log_path, "a")
         else:
-            log_header = ["index", "time", "time_diff", "num_events", "num_changes"]
+            log_header = ["index", "time", "time_diff",
+                          "num_events", "num_changes"]
             self.log_file = open(self.log_path, "a")
             io.write_list(self.log_file, log_header)
 
@@ -207,8 +216,9 @@ class Lines:
     def init_model(self):
         self.model = lstm.LSTM()
         self.loss_fxn = nn.CrossEntropyLoss()
-        self.optimizer = optim.SGD(self.model.parameters(), lr=0.001, momentum=0.9)
-        self.teams_dict, self.statuses_dict = bov.dicts_for_one_hotting()
+        self.optimizer = optim.SGD(
+            self.model.parameters(), lr=0.001, momentum=0.9)
+        self.teams_dict, self.statuses_dict = hot.dicts_for_one_hotting()
         self.running_loss = 0.0
         self.correct = 0
         self.model_log_file = io.init_csv(
@@ -244,7 +254,8 @@ class Lines:
         self.prev_time = self.new_time
 
         if self.keep_open:
-            self.files = write_opened(self.dir, self.files, to_write, verbose=self.verb)
+            self.files = write_opened(
+                self.dir, self.files, to_write, verbose=self.verb)
         else:
             self.files = open_and_write(
                 self.dir, self.files, to_write, verbose=self.verb
@@ -297,10 +308,10 @@ class Lines:
             prev_mls = prevs[16:18]
             cur_mls = v[16:18]
 
-            true_transition = bov.classify_transition(prev_mls, cur_mls)
+            true_transition = analyze.classify_transition(prev_mls, cur_mls)
 
             X = torch.tensor(
-                bov.serialize_row(prevs, self.teams_dict, self.statuses_dict)
+                s.serialize_row(prevs, self.teams_dict, self.statuses_dict)
             )
             self.optimizer.zero_grad()
