+++ sips/macros/sports/nfl.py
