+++ sips/macros/sports/nhl.py
