+++ sips/macros/sports_ref.py
@@ -44,39 +44,6 @@ mlb_teams_short = [
     "WSN",
 ]
 
-mlb_teams_full = [
-    "Arizona Diamondbacks",
-    "Atlanta Braves",
-    "Baltimore Orioles",
-    "Boston Red Sox",
-    "Chicago Cubs",
-    "Chicago White Sox",
-    "Cincinnati Reds",
-    "Cleveland Indians",
-    "Colorado Rockies",
-    "Detroit Tigers",
-    "Houston Astros",
-    "Kansas City Royals",
-    "Los Angeles Angels",
-    "Los Angeles Dodgers",
-    "Miami Marlins",
-    "Milwaukee Brewers",
-    "Minnesota Twins",
-    "New York Mets",
-    "New York Yankees",
-    "Oakland Athletics",
-    "Philadelphia Phillies",
-    "Pittsburgh Pirates",
-    "San Diego Padres",
-    "San Francisco Giants",
-    "Seattle Mariners",
-    "St. Louis Cardinals",
-    "Tampa Bay Rays",
-    "Texas Rangers",
-    "Toronto Blue Jays",
-    "Washington Nationals",
-]
-
 
 mlb_csv_columns = [
     "game_id",
